Use monotonically increasing frame_ids instead of modulo-ed frame IDs
Delete frame_count field from frame_stats


Former-commit-id: fbcdd86f61f0868e029e6e2ffb3aa3e9dcbf7705

+++ src/millipede/dozf.hpp
@@ -52,7 +52,7 @@ public:
 
 private:
     void ZF_time_orthogonal(size_t tag);
-    void precoder(void* mat_input, int frame_id, int sc_id, int offset,
+    void precoder(void* mat_input, size_t frame_id, size_t sc_id, size_t offset,
         bool downlink_mode);
 
     void ZF_freq_orthogonal(size_t tag);
